dbparser: get rid off patterndb-int.h
As we have the broken-down header files, this is not needed anymore.
