--- modules/dbparser/patterndb-int.h
@@ -1,80 +1,0 @@
-/*
- * Copyright (c) 2002-2013 BalaBit IT Ltd, Budapest, Hungary
- * Copyright (c) 1998-2013 Balázs Scheidler
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 as published
- * by the Free Software Foundation, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
- *
- * As an additional exemption you are allowed to compile & link against the
- * OpenSSL libraries as published by the OpenSSL project. See the file
- * COPYING for details.
- *
- * Internal declarations to be used by pdbtool/patterndb internals. Not a
- * public API.
- *
- */
-#ifndef PATTERNDB_INT_H_INCLUDED
-#define PATTERNDB_INT_H_INCLUDED
-
-#include "patterndb.h"
-#include "correllation.h"
-#include "correllation-context.h"
-#include "synthetic-message.h"
-
-typedef struct _PDBLookupParams PDBLookupParams;
-typedef struct _PDBRule PDBRule;
-
-/* This class encapsulates a correllation context, keyed by CorrellationKey, type == PSK_RULE. */
-typedef struct _PDBContext
-{
-  CorrellationContext super;
-  /* back reference to the PatternDB */
-  PatternDB *db;
-  /* back reference to the last rule touching this context */
-  PDBRule *rule;
-  /* timeout timer */
-  TWEntry *timer;
-} PDBContext;
-
-/* This class encapsulates a rate-limit state stored in
-   db->state. */
-typedef struct _PDBRateLimit
-{
-  /* key in the hashtable. NOTE: host/program/pid/session_id are allocated, thus they need to be freed when the structure is freed. */
-  CorrellationKey key;
-  gint buckets;
-  guint64 last_check;
-} PDBRateLimit;
-
-struct _PDBLookupParams
-{
-  LogMessage *msg;
-  NVHandle program_handle;
-  NVHandle message_handle;
-  const gchar *message_string;
-  gssize message_len;
-};
-
-struct _PatternDB
-{
-  GStaticRWLock lock;
-  PDBRuleSet *ruleset;
-  CorrellationState correllation;
-  GHashTable *rate_limits;
-  TimerWheel *timer_wheel;
-  GTimeVal last_tick;
-  PatternDBEmitFunc emit;
-  gpointer emit_data;
-};
-
-#endif
